services/llm_service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
redis>=4.0.0
tenacity>=8.0.0
aiohttp>=3.8.0
python-multipart>=0.0.5
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.5.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
langchain>=0.0.300
numpy>=2.2.0
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.15.0
python-dotenv==1.0.0

# Test-Abhängigkeiten
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.5
pytest-timeout==2.2.0
pytest-xdist==3.5.0
pytest-benchmark==4.0.0
pytest-env==1.1.3
pytest-randomly==3.15.0
pytest-sugar==1.0.0
pytest-html==4.1.1
pytest-metadata==3.1.0
pytest-clarity==1.0.1
pytest-picked==0.5.0
pytest-socket==0.6.0
pytest-freezegun==0.4.2
pytest-responsemock==1.2.0
pytest-httpserver==1.0.8
pytest-mock-resources==2.4.0
pytest-docker==2.0.1
pytest-kafka==0.1.0
pytest-redis==4.0.0
pytest-mongodb==2.1.0
pytest-postgresql==4.1.1
pytest-mysql==4.1.1
pytest-rabbitmq==2.0.0
pytest-elasticsearch==0.4.0
pytest-cassandra==0.1.0
pytest-neo4j==0.1.0
pytest-graphql==0.1.0
pytest-grpc==0.1.0
pytest-websocket==0.1.0
pytest-mqtt==0.1.0
pytest-amqp==0.1.0
pytest-nats==0.1.0 